[Fix]バグ修正

+++ app/controllers/public/favorites_controller.rb
@@ -2,15 +2,16 @@ class Public::FavoritesController < ApplicationController
 
   def create
     post = Post.find(params[:post_id])
-    favorite = current_user.favorites.new(post_id: post.id)
-    favorite.save
+    current_user.favorites.find_or_create_by(post_id: post.id)
+    #favorite = current_user.favorites.new(post_id: post.id)
+    #favorite.save
     redirect_to post_path(post)
   end
 
   def destroy
     post = Post.find(params[:post_id])
     favorite = current_user.favorites.find_by(post_id: post.id)
-    favorite.destroy
+    favorite.destroy if favorite
     redirect_to post_path(post)
   end
